+++ src/main/java/akoletter/devakoletterapi/post/PostListLoad/domain/response/PostListLoadResponse.java
@@ -28,7 +28,7 @@ public class PostListLoadResponse {
     private long unqUsrId;
 
     @Schema(description = "첨부파일 id")
-    private Long fileId;
+    private int fileId;
