+++ src/main/java/akoletter/devakoletterapi/util/File/service/FileService.java
@@ -12,6 +12,6 @@ import java.util.Optional;
 
 public interface FileService {
     List<FileMst> addBoard(FileMst board, List<MultipartFile> files) throws Exception;
-    Optional<FileMst> findBoard(Long fileId);
+    Optional<FileMst> findBoard(int fileId);
     List<FileMst> findBoards();
 }
